save project

+++ app/src/main/java/com/alpha/museum/museum/models/Beacon.java
@@ -4,27 +4,27 @@ import java.io.Serializable;
 
 public class Beacon implements Serializable {
 
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    public void Beacon(int x, int y) {
+    public void Beacon(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 }
